Replace hasSQLitePlugin() disk check with hasMysqlWpConfig() wp-config check
- Move hasMysqlWpConfig() to tools/common/lib/fs-utils so it can be shared
- Replace the old on-disk plugin/db.php/database detection in SiteServer with
  a simple isSqliteSite() that returns !hasMysqlWpConfig(), consistent with
  how the CLI child process determines SQLite vs MySQL
- Remove the now-unused fsExtra import from site-server.ts
